# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JA-CMS API Documentation",
        "description": "Comprehensive API documentation for JA-CMS Content Management System",
        "contact": {
            "email": "support@jejakawan.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/ja/dashboard/admin": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get admin dashboard data",
                "operationId": "3a56307d116e6a972d355d7c54072d1d",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/ja/dashboard/creator": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get creator dashboard data",
                "operationId": "91e61dd843d74a092d8054da544e186f",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/ja/dashboard/viewer": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get viewer dashboard data",
                "operationId": "9557c6c3f410923a7a302cd45b5759c9",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}